Combine Proxy and SlotsProxy
